requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.10
django-environ==0.8.1
django-phonenumber-field==6.0.0
phonenumbers==8.12.40
gunicorn==20.1.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
